% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utility.R
\docType{package}
\name{BayesGP-package}
\alias{BayesGP}
\alias{BayesGP-package}
\title{BayesGP: Efficient Implementation of Gaussian Processes in Bayesian Hierarchical Models}
\description{
Implements Bayesian hierarchical models with flexible Gaussian process priors.
}
\keyword{internal}
