% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utility.R
\name{compute_d_step_sgpsd}
\alias{compute_d_step_sgpsd}
\title{Compute the SD correction factor for sgp}
\usage{
compute_d_step_sgpsd(d, a)
}
\arguments{
\item{d}{A numeric value for the prediction step.}

\item{a}{The frequency parameter of the sgp.}
}
\value{
The correction factor c that should be used to compute the d-step PSD as c*SD.
}
\description{
Compute the SD correction factor for sgp
}
