% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utility.R
\name{compute_weights_precision}
\alias{compute_weights_precision}
\alias{compute_weights_precision,iwp-method}
\title{Constructing the precision matrix given the knot sequence}
\usage{
compute_weights_precision(object)

\S4method{compute_weights_precision}{iwp}(object)
}
\arguments{
\item{object}{An `iwp` smooth-term object.}
}
\value{
A precision matrix of the corresponding basis function, should be diagonal matrix with
size (k-1) by (k-1).
}
\description{
Constructing the precision matrix given the knot sequence
}
