% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utility.R
\name{compute_weights_precision_helper}
\alias{compute_weights_precision_helper}
\title{Constructing the precision matrix given the knot sequence (helper)}
\usage{
compute_weights_precision_helper(x)
}
\arguments{
\item{x}{A vector of knots used to construct the O-spline basis, first knot should be viewed as "0",
the reference starting location. These k knots will define (k-1) basis function in total.}
}
\value{
A precision matrix of the corresponding basis function, should be diagonal matrix with
size (k-1) by (k-1).
}
\description{
Constructing the precision matrix given the knot sequence (helper)
}
\examples{
compute_weights_precision(x = c(0,0.2,0.4,0.6,0.8))
}
