% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utility.R
\name{f_iid}
\alias{f_iid}
\alias{f.iid}
\title{Model-specific `f()` documentation for independent random effects}
\description{
Use `f(group, model = "iid")` for exchangeable random effects over observed
levels of a grouping variable.
}
\section{Required arguments}{

\describe{
  \item{`smoothing_var`}{Grouping variable in `data`. Character, factor, and
  numeric inputs are converted to observed levels.}
  \item{`model`}{Use `model = "iid"`.}
}
}

\section{Optional arguments}{

\describe{
  \item{`sd.prior`}{Prior for the random-effect standard deviation. The
  default is `list(prior = "exp", param = list(u = 1, alpha = 0.5))`.}
}
}

\examples{
f(group, model = "iid")
f(group, model = "iid",
  sd.prior = list(prior = "exp", param = list(u = 0.5, alpha = 0.1))
)

}
\seealso{
\code{\link{f}}, \code{\link{supported_models}}
}
