% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utility.R
\name{get_default_option_list_MCMC}
\alias{get_default_option_list_MCMC}
\title{Construct default MCMC options}
\usage{
get_default_option_list_MCMC(option_list = list())
}
\arguments{
\item{option_list}{Optional named list overriding default MCMC options.}
}
\value{
A named list of MCMC options.
}
\description{
Construct default MCMC options
}
