% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_post_fit.R
\name{hyper_summary}
\alias{hyper_summary}
\title{Obtain the posterior summary of a hyper-parameter in the fitted model}
\usage{
hyper_summary(object, component = NULL, h = NULL, theta_logprior = NULL)
}
\arguments{
\item{object}{The fitted object from the function `model_fit`.}

\item{component}{The component of the hyper-parameter that you want to show. By default this is `NULL`, indicating the family.sd is of interest.}

\item{h}{For PSD, the unit of predictive step to consider, by default is set to `NULL`, indicating the result is using the same `h` as in the model fitting.}

\item{theta_logprior}{The log prior function used on the selected hyper-parameter. By default is `NULL`, and the default Exponential prior will be used.}
}
\description{
Obtain the posterior summary of a hyper-parameter in the fitted model
}
