% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_post_fit.R
\name{para_density}
\alias{para_density}
\title{Obtain the posterior and prior density of all the parameters in the fitted model}
\usage{
para_density(object)
}
\arguments{
\item{object}{The fitted object from the function `model_fit`.}
}
\description{
Obtain the posterior and prior density of all the parameters in the fitted model
}
