% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PEN_death}
\alias{PEN_death}
\title{The monthly all-cause mortality for male with age less than 40 in Pennsylvania.}
\format{
'PEN_death'
A data frame with 299 rows of observations.
}
\usage{
PEN_death
}
\description{
The monthly all-cause mortality for male with age less than 40 in Pennsylvania.
}
\keyword{datasets}
