% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sample_fixed_effect}
\alias{sample_fixed_effect}
\title{Extract the posterior samples from the fitted model for the target fixed variables.}
\usage{
sample_fixed_effect(model_fit, variables, M)
}
\arguments{
\item{model_fit}{The result from model_fit().}

\item{variables}{A vector of names of the target fixed variables to sample.}

\item{M}{The number of posterior samples to take.}
}
\description{
Extract the posterior samples from the fitted model for the target fixed variables.
}
