% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_post_fit.R
\name{sample_sd}
\alias{sample_sd}
\title{Extract the posterior samples from the fitted model for the target standard deviation.}
\usage{
sample_sd(model_fit, variables, M = NULL, h = NULL)
}
\arguments{
\item{model_fit}{The result from model_fit().}

\item{variables}{A vector of names of the target standard deviation to sample.}

\item{M}{The number of samples to draw from the posterior distribution.}

\item{h}{The unit of predictive step to consider, by default is set to `NULL`, indicating the result is using the same `h` as in the model fitting.}
}
\value{
A matrix with columns being the posterior samples of the target sd variables.
}
\description{
Extract the posterior samples from the fitted model for the target standard deviation.
}
