% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim1data}
\alias{sim1data}
\title{A simulated dataset from the case-crossover model.}
\format{
'sim1data'
A data frame with 3596 rows and 6 columns.
}
\usage{
sim1data
}
\description{
A simulated dataset from the case-crossover model.
}
\keyword{datasets}
