% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_weights_precision}
\alias{compute_weights_precision}
\title{Constructing the precision matrix given the knot sequence}
\usage{
compute_weights_precision(object)
}
\arguments{
\item{x}{A vector of knots used to construct the O-spline basis, first knot should be viewed as "0",
the reference starting location. These k knots will define (k-1) basis function in total.}
}
\value{
A precision matrix of the corresponding basis function, should be diagonal matrix with
size (k-1) by (k-1).
}
\description{
Constructing the precision matrix given the knot sequence
}
