% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utility.R
\name{f}
\alias{f}
\title{Function defined to enhance the usability for users on IDEs.}
\usage{
f(
  smoothing_var,
  model = "iid",
  sd.prior = NULL,
  boundary.prior = NULL,
  initial_location = c("middle", "left", "right"),
  ...
)
}
\description{
Function defined to enhance the usability for users on IDEs.
}
