% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_utility.R
\name{get_default_option_list_MCMC}
\alias{get_default_option_list_MCMC}
\title{Get default options for MCMC implementation}
\usage{
get_default_option_list_MCMC(option_list = list())
}
\arguments{
\item{option_list}{A list of options to be passed to the MCMC. If some options are missing, the function will use default values.}
}
\value{
A list containing the complete set of options with defaults where necessary.
}
\description{
This function takes an optional list of options and fills in any missing values with a set of default MCMC options.
}
\examples{
# Example: Get the default option list
options <- get_default_option_list_MCMC()
print(options)

}
