% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_post_fit.R
\name{post_table}
\alias{post_table}
\title{Obtain the posterior summary table for all the parameters in the fitted model}
\usage{
post_table(object, quantiles = c(0.025, 0.975))
}
\arguments{
\item{quantiles}{The specified quantile to display the posterior summary, default is c(0.025, 0.975).}
}
\description{
Obtain the posterior summary table for all the parameters in the fitted model
}
